requirements.txt
# src/main.py
pyside6==6.10.1

ollama==0.6.1
pillow==12.0.0
pymupdf==1.26.6
markdown==3.10